Unified and sorted pages and other resources.
The Grid Framework pages are now all in their own directory, images
have been split up where the Grid Framework images are in the Grid
Framework directory and the button graphics are globally available. The
header bar CSS is now its own file as well. Links are now relative to
their current locations.

The part that's still annoying me is that the links in the nav bar have
to be adjusted manually.